Specify that mnesia:start/0 is asynchronous
I think this will save a lot of time for future newcomers, like myself. I may have been searching in wrong places and didn't do things just right while investigating, but it took me too long to find out that I needed to use the `mnesia:wait_for_tables/2` after Mnesia start to do table transformations. I have just copy-pasted the text from the Mnesia guide and IMHO it's more important than the rest of the stuff in the function description, so I put it at the beginning. If you disagree about that or with the wording, feel free to change it, but please at least mention that it's async and the `mnesia:wait_for_tables/2` function.